Changed pressure threshold for inair identifier.
Changed pressure threshold from 0.16 to 0.10 dbar for identifying true
in'air obs when optode is mounted on a stick (no dedicated in-air
sequence).  Settling on the best method for identifying in-air obs
within traj files, across all float platforms, is still a work in
progress.  User may want to modify this mfile to best represent
extraction of inair obs for his/her particular float type.