Replaced phcalc_jp.m with phcalc.m
Replaced phcalc_jp.m with phcalc.m.  The latter is the most up-to-date
and is the code referenced
in the Argo documentation (https://doi.org/10.13155/57195).  The two
pieces of code are nearly identical; differences in calculated pH from
the two pieces of code are on the order of ~0.1 millipH units and can be
attributed to differences in format precision of constants used in
calculations.