Added fx to assist in populating Argo meta fields
Added function written to assist AOML with the population of key meta
file fields for BGC parameter data.  This code piece was written in line
with the MBARI processing system and is not intended for direct use in
external systems (modifications will be required) but can serve as an
example for field specification.