Modification to Warning msg re Nans.
Modification to the warning message alerting user to high percentages of
NaNs in the BGC data series.  Too many NaNs may indicate sensor or float
issues that require further attention from the DMQC operator.  However,
this is not always the case.  Wording was modified to be less absolute,
based on user feedback.  Changepoint detection in the prescence of
varying levels of NaNs may be revisited, and this section of the code
enhanced at a future date.